[Feat] 중복 기능 BaseSearchVC로 통합
